workitems: dynamically measure scheduling overhead & remove TUNING_ADJUSTMENT
This continuously measures the scheduling overhead, by how much an item
was scheduled too late. This value is then used to adjust the sleep time,
so that scheduling becomes more accurate.

This has been tested on an Intel i7 CPU and a RPi. It works well in both
cases, however on RPi, scheduling is much more irregular. In fact, when
using too aggressive params (or with the previous TUNING_ADJUSTMENT),
CPU usage increases drastically and scheduling starts to perform worse.